:truck:rename: assets 이미지,아이콘 폴더 분리 및 import문 수정
