+++ src/assets/images/main-hot-item.png
