+++ src/styles/AddItemFileInput.css
@@ -23,7 +23,7 @@
   border-radius: 12px;
   background-color: #f3f4f6;
   border: none;
-  background-image: url("../assets/file-input-button.svg");
+  background-image: url("../assets/images/file-input-button.svg");
   background-repeat: no-repeat;
   background-position: center;
   cursor: pointer;
@@ -42,7 +42,7 @@
 }
 
 .preview-delete-btn {
-  background-image: url("../assets/img-delete-btn.svg");
+  background-image: url("../assets/icons/img-delete-btn.svg");
   background-repeat: no-repeat;
   background-position: center;
   background-color: transparent;
