+++ src/styles/AddItemTagsInput.css
@@ -40,7 +40,7 @@
 }
 
 .tag-delete-button {
-  background-image: url("../assets/icon-delete-tag-btn.svg");
+  background-image: url("../assets/icons/icon-delete-tag-btn.svg");
   background-repeat: no-repeat;
   width: 22px;
   height: 24px;
