+++ src/styles/AllItemList.css
@@ -66,8 +66,8 @@
   width: 325px;
   height: 42px;
   padding: 9px, 20px, 16px;
-  background: #f3f4f6 url(../assets/item-search-icon.svg) no-repeat left 16px
-    center;
+  background: #f3f4f6 url(../assets/icons/item-search-icon.svg) no-repeat left
+    16px center;
   background-size: 15px;
 }
 
@@ -87,7 +87,8 @@ input.all-item-header-search::placeholder {
   height: 42px;
   padding: 9px 20px;
   appearance: none;
-  background: url(../assets/arrow-down-icon.svg) no-repeat right 17px center;
+  background: url(../assets/icons/arrow-down-icon.svg) no-repeat right 17px
+    center;
 }
 .select-box:focus {
   outline: none;
@@ -173,6 +174,6 @@ input.all-item-header-search::placeholder {
   } */
   .select-box {
     width: 42px;
-    background: url(../assets/arrow-down-icon.svg) no-repeat center;
+    background: url(../assets/icons/arrow-down-icon.svg) no-repeat center;
   }
 }
