+++ src/styles/Pagination.css
@@ -31,18 +31,18 @@
 }
 
 .disabled-prev {
-  background-image: url("../assets/arrow-left-disabled.svg");
+  background-image: url("../assets/icons/arrow-left-disabled.svg");
   cursor: default;
 }
 
 .active-prev {
-  background-image: url("../assets/arrow-left.svg");
+  background-image: url("../assets/icons/arrow-left.svg");
 }
 .disabled-next {
-  background-image: url("../assets/arrow-right-disabled.svg");
+  background-image: url("../assets/icons/arrow-right-disabled.svg");
   cursor: default;
 }
 
 .active-next {
-  background-image: url("../assets/arrow-right.svg");
+  background-image: url("../assets/icons/arrow-right.svg");
 }
